pyproject.toml
[tool.poetry]
name = "warden"
version = "0.1.0"
description = ""
authors = ["Giannis Papaioannou <giannispapcod7@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.25.1"
python-decouple = "^3.3"
rich = "^9.4.0"

[tool.poetry.dev-dependencies]
isort = "^5.6.4"
black = "^20.8b1"
flake8 = "^3.8.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
